+++ DuoArtifacts/DuoUpsideChoice.cs
@@ -1,0 +1,3 @@
+﻿namespace Shockah.DuoArtifacts;
+
+internal sealed class DuoUpsideChoice : Choice;