+++ DuoArtifacts/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * Removed Perfect Shield from the Energy Condenser (Dizzy-Peri duo) artifact's tooltip, as it is no longer relevant.
 * Shortened the tooltip header of duo artifacts, making most (all?) of them fit in one line.
+* Added Custom Run Options integration.
 
 ## 1.8.2
 Released 14 September 2025.